abstract adaptive cards lookup function
